+++ flowman-dsl/src/test/scala/com/dimajix/flowman/dsl/ExampleSpec.scala
@@ -24,7 +24,6 @@ class ExampleSpec extends AnyFlatSpec with Matchers with LocalSparkSession {
             .build()
 
         val context = session.getContext(project)
-        val executor = session.execution
         val runner = session.runner
 
         val job = context.getJob(JobIdentifier("test"))